refactor: move lua logic to separate struct
