+++ src/ui/windows/loading/check_addons.rs
@@ -43,7 +43,7 @@ fn get_addon_download(
     if is_addon_enabled(enabled_addons, addon, group) {
         let addon_path = addon.get_installation_path(&group.name, &game.manifest.game_name, edition)?;
 
-        let installed = game.is_addon_installed(
+        let installed = game.driver.is_addon_installed(
             &group.name,
             &addon.name,
             &addon_path.to_string_lossy(),
@@ -60,7 +60,7 @@ fn get_addon_download(
             return Ok(Some(entry));
         }
 
-        let diff = game.get_addon_diff(
+        let diff = game.driver.get_addon_diff(
             &group.name,
             &addon.name,
             &addon_path.to_string_lossy(),
@@ -85,7 +85,7 @@ pub fn get_game_addons_downloads(
 ) -> anyhow::Result<Vec<AddonsListEntry>> {
     let mut result = Vec::new();
 
-    for group in game.get_addons_list(edition)? {
+    for group in game.driver.get_addons_list(edition)? {
         for addon in &group.addons {
             if let Some(addon) = get_addon_download(game_info, game, edition, enabled_addons, addon, &group)? {
                 result.push(addon);
@@ -111,8 +111,8 @@ pub fn get_download(pool: &rusty_pool::ThreadPool) -> anyhow::Result<Vec<AddonsL
         tasks.push(pool.evaluate(move || -> anyhow::Result<Vec<AddonsListEntry>> {
             let mut addons = Vec::new();
 
-            for edition in game.get_game_editions_list()? {
-                let installed = game.is_game_installed(
+            for edition in game.driver.get_game_editions_list()? {
+                let installed = game.driver.is_game_installed(
                     &settings.paths[&edition.name].game.to_string_lossy(),
                     &edition.name
                 )?;
@@ -125,7 +125,7 @@ pub fn get_download(pool: &rusty_pool::ThreadPool) -> anyhow::Result<Vec<AddonsL
                         name: game.manifest.game_name.clone(),
                         title: game.manifest.game_title.clone(),
                         developer: game.manifest.game_developer.clone(),
-                        picture_uri: game.get_card_picture(&edition.name)?,
+                        picture_uri: game.driver.get_card_picture(&edition.name)?,
                         edition: edition.name.clone()
                     };
